chore: implement pre post patch
