+++ packages/vite/package.json
@@ -32,8 +32,7 @@
     "client.d.ts",
     "index.cjs",
     "src/client",
-    "types",
-    "client/types.d.ts"
+    "types"
   ],
   "engines": {
     "node": "^14.18.0 || >=16.0.0"
@@ -51,13 +50,12 @@
     "dev": "rimraf dist && pnpm run build-bundle -w",
     "build": "rimraf dist && run-s build-bundle build-types",
     "build-bundle": "rollup --config rollup.config.ts --configPlugin typescript",
-    "build-types": "run-p build-node-types build-client-types",
-    "build-node-types": "run-s build-node-types-temp build-node-types-patch build-node-types-roll build-node-types-check",
-    "build-node-types-temp": "tsc --emitDeclarationOnly --outDir temp/node -p src/node",
-    "build-node-types-patch": "tsx scripts/patchTypes.ts",
-    "build-node-types-roll": "api-extractor run && rimraf temp",
-    "build-node-types-check": "tsc --project tsconfig.check.json",
-    "build-client-types": "api-extractor run -c api-extractor.client.json",
+    "build-types": "run-s build-types-temp build-types-pre-patch build-types-roll build-types-post-patch build-types-check",
+    "build-types-temp": "tsc --emitDeclarationOnly --outDir temp/node -p src/node",
+    "build-types-pre-patch": "tsx scripts/prePatchTypes.ts",
+    "build-types-roll": "api-extractor run && rimraf temp",
+    "build-types-post-patch": "tsx scripts/postPatchTypes.ts",
+    "build-types-check": "tsc --project tsconfig.check.json",
     "lint": "eslint --cache --ext .ts src/**",
     "format": "prettier --write --cache --parser typescript \"src/**/*.ts\"",
     "prepublishOnly": "npm run build"
